waveform generation

+++ app/uploaders/song_mixaudio_uploader.rb
@@ -1,7 +1,15 @@
 class SongMixaudioUploader < CarrierWave::Uploader::Base
   include ::CarrierWave::Backgrounder::Delay
+  before :cache, :save_duration
 
   process encode_audio: [:m4a]
+  process encode_audio: [:mp3, false]
+  process :generate_waveform
+
+  def save_duration(file)
+    file = ::FFMPEG::Movie.new(file.path)
+    model.duration = file.duration
+  end
 
   # Override the directory where uploaded files will be stored.
   def store_dir
@@ -35,10 +43,14 @@ class SongMixaudioUploader < CarrierWave::Uploader::Base
     model.instance_variable_get(var) or model.instance_variable_set(var, SecureRandom.hex(length/2))
   end
 
-  def encode_audio(format='m4a')
+  def encode_audio(format='m4a', overwrite=true)
     directory = File.dirname(current_path)
     tmpfile = File.join(directory, 'tmpfile')
-    File.rename(current_path, tmpfile)
+    if overwrite
+      FileUtils.mv(current_path, tmpfile)
+    else
+      FileUtils.cp(current_path, tmpfile)
+    end
 
     file = ::FFMPEG::Movie.new(tmpfile)
     new_name = File.basename(current_path, '.*') + '.' + format.to_s
@@ -47,9 +59,46 @@ class SongMixaudioUploader < CarrierWave::Uploader::Base
 
     file.transcode(encoded_file)
 
-    self.filename[-current_extension.size..-1] = format.to_s
-    self.file.file[-current_extension.size..-1] = format.to_s
+    if overwrite
+      self.filename[-current_extension.size..-1] = format.to_s
+      self.file.file[-current_extension.size..-1] = format.to_s
 
-    File.delete(tmpfile)
+      File.delete(tmpfile)
+    end
+  end
+
+  def generate_waveform
+    directory = File.dirname(current_path)
+
+    waveform_line = Cocaine::CommandLine.new('audiowaveform', '-i :input -s :start -e :end -w :width -h :height -o :output --no-axis-labels --waveform-color :waveform_color --background-color :background_color')
+    interpolations = {
+        input: File.join(directory, File.basename(current_path, '.*') + '.mp3'),
+        start: 0,
+        end: model.duration,
+        width: 640,
+        height: 80,
+        output: File.join(directory, 'waveform.png'),
+        waveform_color: 'aaaaaa',
+        background_color: 'ffffff00'
+    }
+
+    begin
+      waveform_line.run(interpolations)
+      self.model.waveform = File.open(interpolations[:output])
+    rescue Cocaine::ExitStatusError => e
+      puts "error while running command #{waveform_line.command(interpolations)}: #{e}"
+    end
+
+    waveform_data_line = Cocaine::CommandLine.new('audiowaveform', '-i :input -o :output')
+    interpolations = {
+        input: File.join(directory, File.basename(current_path, '.*') + '.mp3'),
+        output: File.join(directory, 'waveform.json')
+    }
+    begin
+      waveform_data_line.run(interpolations)
+      self.model.waveform_data = File.open(interpolations[:output])
+    rescue Cocaine::ExitStatusError => e
+      puts "error while running command #{waveform_data_line.command}: #{e}"
+    end
   end
 end
